requirements.txt
PyYAML>=6.0.1
flake8==6.1.0
black==23.9.1
usort==1.0.7
ufmt==2.3.0
pre-commit==3.5.0
prettier==0.0.7
